Fix Cart

+++ hawa s/app/services/cartItem.ts
@@ -1,23 +1,51 @@
 import { PrismaClient } from '@prisma/client'
+import { decrementShoppingSessionTotal, incrementShoppingSessionTotal } from './shoppingSession'
+import { decrementProductInventoryQuantity, getProduct, incrementProductInventoryQuantity } from './product'
 
 const prisma = new PrismaClient()
 
 export const createCartItem = async (sessionId: number, productId: number, quantity: number) => {
-    return await prisma.cartItem.create({
+    const existingCartItem = await getCartItemForSessionAndProduct(sessionId, productId)
+
+    if (existingCartItem) {
+        const updateCartItem = await incrementCartItemQuantity(existingCartItem.id, quantity)
+
+        await incrementShoppingSessionTotal(sessionId, Number(updateCartItem.product.price) * quantity)
+
+        return updateCartItem
+    }
+
+    const product = await getProduct(productId)
+
+    if (!product) {
+        throw new Error('Producto no encontrado')
+    }
+
+    if (product.inventory.quantity < quantity) {
+        throw new Error('Existencias insuficiente')
+    }
+
+    const newCartItem = await prisma.cartItem.create({
         data: {
-            session: {
-                connect: {
-                    id: sessionId
-                }
-            },
-            product: {
-                connect: {
-                    id: productId
-                }
-            },
+            sessionId,
+            productId,
             quantity
+        },
+        include: {
+            session: true,
+            product: {
+                include: {
+                    discount: true
+                }
+            },
         }
     })
+
+    await incrementShoppingSessionTotal(sessionId, Number(newCartItem.product.price) * (1 - Number(newCartItem.product.discount?.discountPercent)) * quantity)
+
+    await decrementProductInventoryQuantity(productId, quantity)
+
+    return newCartItem
 }
 
 export const getCartItems = async () => {
@@ -51,6 +79,26 @@ export const getCartItem = async (id: number) => {
     })
 }
 
+export const getCartItemForSessionAndProduct = async (sessionId: number, productId: number) => {
+    return await prisma.cartItem.findUnique({
+        where: {
+            sessionId_productId: {
+                sessionId,
+                productId
+            }
+        },
+        include: {
+            session: true, 
+            product: {
+                include: {
+                    discount: true,
+                    inventory: true
+                }
+            }
+        }
+    })
+}
+
 export const updateCartItem = async (id: number, sessionId: number, productId: number, quantity: number) => {
     return await prisma.cartItem.update({
         where: {
@@ -70,12 +118,59 @@ export const updateCartItem = async (id: number, sessionId: number, productId: n
             quantity
         }
     })
+
+}
+
+export const incrementCartItemQuantity = async (id: number, quantity: number) => {
+    return await prisma.cartItem.update({
+        where: {
+            id
+        },
+        data: {
+            quantity: {
+                increment: quantity
+            }
+        },
+        include: {
+            session: true,
+            product: true
+        }
+    })
+}
+
+export const decrementCartItemQuantity = async (id: number, quantity: number) => {
+    return await prisma.cartItem.update({
+        where: {
+            id
+        },
+        data: {
+            quantity: {
+                decrement: quantity
+            }
+        },
+        include: {
+            session: true,
+            product: true
+        }
+    })
 }
 
 export const deleteCartItem = async (id: number) => {
-    return await prisma.cartItem.delete({
+    const cartItem = await getCartItem(id)
+
+    if (!cartItem) {
+        throw new Error('Producto no encontrado en el carrito')
+    }
+
+    await prisma.cartItem.delete({
         where: {
             id
         }
     })
+
+    await decrementShoppingSessionTotal(cartItem.sessionId, Number(cartItem.product.price) * (1 - Number(cartItem.product.discount?.discountPercent)) * cartItem.quantity)
+
+    await incrementProductInventoryQuantity(cartItem.productId, cartItem.quantity)
+
+    return { success: true }
 }
